+++ internal/domain/repository/user.go
@@ -8,9 +8,9 @@ import (
 
 type UserRepository interface {
 	Create(ctx context.Context, user *entity.User) error
-	GetByID(ctx context.Context, id uint) (*entity.User, error)
-	GetWithTargetAccounts(ctx context.Context, id uint) (*entity.User, error)
+	GetByID(ctx context.Context, id int64) (*entity.User, error)
+	GetWithTargetAccounts(ctx context.Context, id int64) (*entity.User, error)
 	Update(ctx context.Context, user *entity.User) error
-	Delete(ctx context.Context, id uint) error
-	Exists(ctx context.Context, id uint) (bool, error)
+	Delete(ctx context.Context, id int64) error
+	Exists(ctx context.Context, id int64) (bool, error)
 }
